auto add to span as well
