+++ nucliadb_telemetry/nucliadb_telemetry/utils.py
@@ -19,13 +19,12 @@
 
 import asyncio
 import os
-from typing import Dict, Optional, Sequence, Union
+from typing import Dict, Optional
 
 from opentelemetry.propagate import set_global_textmap
 from opentelemetry.propagators.b3 import B3MultiFormat
 from opentelemetry.sdk.resources import SERVICE_NAME  # type: ignore
 from opentelemetry.sdk.resources import Resource  # type: ignore
-from opentelemetry.trace import get_current_span
 
 from nucliadb_telemetry.batch_span import BatchSpanProcessor
 from nucliadb_telemetry.jaeger import JaegerExporterAsync
@@ -35,6 +34,10 @@ from nucliadb_telemetry.tracerprovider import (
     AsyncMultiSpanProcessor,
     AsyncTracerProvider,
 )
+
+from .context import set_info_on_span  # noqa: F401
+
+set_info_on_span  # b/w compatible import
 
 GLOBAL_PROVIDER: Dict[str, AsyncTracerProvider] = {}
 
@@ -128,23 +131,3 @@ async def setup_telemetry(service_name: str) -> Optional[AsyncTracerProvider]:
         except ImportError:
             pass
     return tracer_provider
-
-
-def set_info_on_span(
-    headers: Dict[
-        str,
-        Union[
-            str,
-            bool,
-            int,
-            float,
-            Sequence[str],
-            Sequence[bool],
-            Sequence[int],
-            Sequence[float],
-        ],
-    ]
-):
-    if telemetry_settings.jaeger_enabled:
-        span = get_current_span()
-        span.set_attributes(headers)
